Began work on ID Card Functionality
Imported the ID card models

The model slightly confused me since it came with a camera (and lighting point?) wasn't sure if those should be removed, and if not if I should move the rest of the model whenever I move the card around.

Put the XR interactable hand grab point at the middle bottom of the card.

Added attach points in the middle of each object- tried adding 2 so you can add it from each side but it wouldn't let me. Decided it wasn't critical.

The resizing of the cards was of factor 5 (set to 0.2 of original size, we constraints of original proportion.

There is a problem where the tutorial for socket item says to put "Is Trigger" as true for the box colider of the item you are trying to make attachable, but if I do that then they fall through the floor...

Okay so I realized I maty need two seperate colliders for each card, one for the floor and one for other cards. I ran out of time working on it for today so this will be something I continue tommorow.