Added Room Layout, Teleportation Areas, and spawning in shake effect
Changes made:
Added in the new layout
Changed original plane visibility due to clipping
Added 6 teleportation areas onto scene (had to re-add them in because we loaded a new scene with the textures)
Added mesh colliders onto every element in scene (NOTE: players can climb the rocks- I think it's fun and we shouldn't block them, but be aware)
Moved spawn location and menu location (confirm good?)
Loaded new textures for cave and rocks
Loaded normal for cave (NOTE: there seems to be an issue with the lighting inside unity making the normal (the squigly lines across the walls and ceiling) look weird. Aadi messaged caroline about it and it will hopefully be fixed tommorow)

I was waiting for Caroline to finish with the new cave textures and decided to try adding a small feature that makes you shake when you spawn in as if the cave is shaking an collapsing

Added for earthquake effect:
New script "CamerShake" is scripts folder in asset- in it there are configurations for intensity and length (though must change in the prefab as well as it doesn't automatically update I think?)
I kept the shake on low settings and did not include rotation shake because that can be disorienting in VR appearntly
Added a royalty free earthquake rumble sound effect into a new folder called "audio" in assests
Created an audio component inside the scene and attached the sound
Attached the camera shake script to the camera offset

I had a thing where i was super tall in the scene, I noticed that the camera offset was had a +1.26 y position transoformation- I am unsure if it had it before or I accidenatly moved it, but I reset it to 0, and moved up the XR origin. I think right now it is fine, though the player is a bit short maybe? just check inside VR that you feel the player is in a good height.

Feel free to discard of the shake mechanic by removing the script from camera offset. If you decided you like it we can even make it a reccuring event in random times, though I think we should only have it in the begining because it's a bit disorienting.

Feel free to make the shake shorter and less intense (or more) if you so choose.

Again I just thought it'd be cool, but you can discard it.

Also the sound effect is a bit too long- but I didn't find a shorter one- you can change it if you'd like